requirements.txt
streamlit>=1.36,<2.0
google-generativeai>=0.8.0
Pillow>=10.3.0
requests>=2.31.0

# PDF support (optional)
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Export functionality (optional)
pandas>=2.0.0
openpyxl>=3.1.0

# Additional utilities (optional)
# pillow-heif>=0.18.0
# python-magic>=0.4.27
# ruff>=0.5.0
# mypy>=1.10.0
